python/celery-test/celery-test-01/test01/requirements.txt
amqp==2.3.2
astroid==2.0.4
Babel==2.6.0
barcodenumber==0.2.1
billiard==3.5.0.4
celery==4.2.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
confluent-kafka==0.11.6
Django==2.1.3
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-timezone-field==3.0
flower==0.9.2
future==0.16.0
idna==2.7
isort==4.3.4
kafka==1.3.5
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pbr==5.1.1
pylint==2.1.1
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.7
redis==2.10.6
requests==2.20.1
sh==1.12.14
six==1.11.0
SQLAlchemy==1.2.14
tornado==5.1.1
typed-ast==1.1.0
urllib3==1.24.1
vine==1.1.4
wmi-client-wrapper-py3==2018.3
Wrapper==1.1.0b1
wrapt==1.10.11
